pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-bq-bot"
version = "0.1.0"
description = "Discord bot for querying and monitoring a BigQuery anomaly-detection table"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0",
    "google-cloud-bigquery>=3.17.0",
    "google-genai>=1.5.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.8.0",
    "tabulate>=0.9.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
